Rename download-scenes to download
