+++ cli/index.js
@@ -3,4 +3,4 @@ exports['find-mosaics'] = require('./find-mosaics');
 exports['find-quads'] = require('./find-quads');
 exports['find-scenes'] = require('./find-scenes');
 exports['find-workspaces'] = require('./find-workspaces');
-exports['download-scenes'] = require('./download-scenes');
+exports.download = require('./download');
